debrief/requirements.txt
fastapi
uvicorn
sqlalchemy
openai-whisper
langchain-groq
python-multipart
python-dotenv
httpx